+++ client/.dockerignore
