+++ client/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "client",
+  "version": "1.0.0",
+  "main": "app.js",
+  "devDependencies": {},
+  "author": "Rylkov Andrey <fellainthewagon@gmail.com>",
+  "license": "ISC",
+  "description": "Clinic",
+  "scripts": {
+    "start": "node app.js"
+  },
+  "dependencies": {
+    "express": "^4.17.1"
+  }
+}
