+++ client/public/config/config.js
@@ -1,0 +1,5 @@
+export default {
+  protocol: "http",
+  host: "localhost",
+  port: 3000,
+};
