--- project/client/package.json
@@ -1,15 +1,0 @@
-{
-  "scripts": {
-    "start": "http-server -a localhost -p 8000 -c-1"
-  },
-  "dependencies": {
-    "express": "^4.17.1"
-  },
-  "name": "client",
-  "version": "1.0.0",
-  "main": "app.js",
-  "devDependencies": {},
-  "author": "",
-  "license": "ISC",
-  "description": ""
-}
