+++ server/.dockerignore
@@ -1,5 +1,4 @@
 node_modules
 npm-debug.log
 Dockerfile
-__tests__
 .dockerignore
