+++ server/package.json
@@ -1,7 +1,7 @@
 {
   "name": "server",
   "version": "1.0.0",
-  "description": "Doctor's cabinete",
+  "description": "Clinic",
   "main": "index.js",
   "scripts": {
     "dev": "nodemon index",
@@ -38,6 +38,7 @@
   "dependencies": {
     "ajv": "^8.2.0",
     "ajv-formats": "^2.0.2",
+    "config": "^3.3.6",
     "express": "^4.17.1",
     "morgan": "^1.10.0",
     "swagger-jsdoc": "^6.1.0",
