+++ server/src/components/resolutions/resolutionsService.js
